# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: AgroGuard Limited - Agricultural IoT Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /farmers:
    get:
      operationId: listFarmers
      tags: [farmers]
      summary: List all farmers (scoped by role for field officers)
      responses:
        "200":
          description: List of farmers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farmer"
    post:
      operationId: createFarmer
      tags: [farmers]
      summary: Create a new farmer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmerInput"
      responses:
        "201":
          description: Farmer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmerCreateResult"
        "400":
          description: Invalid input
  /farmers/{id}:
    get:
      operationId: getFarmer
      tags: [farmers]
      summary: Get farmer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Farmer found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
        "404":
          description: Farmer not found
    patch:
      operationId: updateFarmer
      tags: [farmers]
      summary: Update farmer details (admin can assign field officer)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmerUpdate"
      responses:
        "200":
          description: Farmer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farmer"
        "404":
          description: Farmer not found
    delete:
      operationId: deleteFarmer
      tags: [farmers]
      summary: Delete farmer (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Farmer deleted
        "404":
          description: Farmer not found
  /farmers/{id}/devices:
    get:
      operationId: getFarmerDevices
      tags: [farmers]
      summary: Get all devices assigned to a farmer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"

  # ── DEVICES ──────────────────────────────────────────────────────────────────
  /devices:
    get:
      operationId: listDevices
      tags: [devices]
      summary: List IoT devices (scoped by role for field officers)
      responses:
        "200":
          description: List of devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
    post:
      operationId: createDevice
      tags: [devices]
      summary: Register a new IoT device
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceInput"
      responses:
        "201":
          description: Device registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "400":
          description: Invalid input
  /devices/{id}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get device by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Device not found
    patch:
      operationId: updateDevice
      tags: [devices]
      summary: Update device info or assign to farmer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeviceUpdate"
      responses:
        "200":
          description: Device updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Device not found
    delete:
      operationId: deleteDevice
      tags: [devices]
      summary: Remove device (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Device deleted
        "404":
          description: Device not found
  /devices/{id}/readings:
    get:
      operationId: getDeviceReadings
      tags: [devices]
      summary: Get recent readings for a device (last 50)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of readings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SensorReading"

  # ── SENSOR READINGS ──────────────────────────────────────────────────────────
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List all farm alerts
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create a farm alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertInput"
      responses:
        "201":
          description: Alert created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}/resolve:
    patch:
      operationId: resolveAlert
      tags: [alerts]
      summary: Mark alert as resolved
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Alert not found

  # ── AI RECOMMENDATIONS ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    Farmer:
      type: object
      required: [id, name, email, phone, location, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        location:
          type: string
        farmName:
          type: ["string", "null"]
        farmSizeHectares:
          type: ["number", "null"]
        cropTypes:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, pending]
        whatsappNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        fieldOfficerId:
          type: ["integer", "null"]
          description: ID of the field officer assigned to manage this farmer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FarmerCreateResult:
      type: object
      required: [id, name, email, phone, location, status, createdAt, tempPassword]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        location:
          type: string
        farmName:
          type: ["string", "null"]
        farmSizeHectares:
          type: ["number", "null"]
        cropTypes:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, pending]
        whatsappNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        fieldOfficerId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        tempPassword:
          type: string
          description: One-time temporary password shown only at creation; the farmer must change it on first login.
    FarmerInput:
      type: object
      required: [name, email, phone, location]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
        location:
          type: string
        farmName:
          type: string
        farmSizeHectares:
          type: number
        cropTypes:
          type: string
        whatsappNumber:
          type: string
        notes:
          type: string
        fieldOfficerId:
          type: integer
          description: Assign a field officer at creation time
    FarmerUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        location:
          type: string
        farmName:
          type: string
        farmSizeHectares:
          type: number
        cropTypes:
          type: string
        status:
          type: string
          enum: [active, inactive, pending]
        whatsappNumber:
          type: string
        notes:
          type: string
        fieldOfficerId:
          type: ["integer", "null"]
          description: Assign or unassign a field officer (null to unassign)

    # ── DEVICE ──────────────────────────────────────────────────────────────────
    Device:
      type: object
      required: [id, deviceId, name, status, createdAt]
      properties:
        id:
          type: integer
        deviceId:
          type: string
          description: Unique hardware identifier (ESP32 chip ID / auto-generated UUID)
        name:
          type: string
        farmerId:
          type: ["integer", "null"]
          description: Assigned farmer ID, null if unassigned
        location:
          type: ["string", "null"]
        status:
          type: string
          enum: [online, offline, maintenance]
        lastSeenAt:
          type: ["string", "null"]
          format: date-time
        firmwareVersion:
          type: ["string", "null"]
        batteryLevel:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DeviceInput:
      type: object
      required: [deviceId, name]
      properties:
        deviceId:
          type: string
          minLength: 1
          description: Unique hardware ID (use the auto-generated AGR-XXXX-XXXX format or ESP32 MAC)
        name:
          type: string
          minLength: 1
        farmerId:
          type: integer
        location:
          type: string
        firmwareVersion:
          type: string
        notes:
          type: string
    DeviceUpdate:
      type: object
      properties:
        name:
          type: string
        farmerId:
          type: ["integer", "null"]
        location:
          type: string
        status:
          type: string
          enum: [online, offline, maintenance]
        firmwareVersion:
          type: string
        batteryLevel:
          type: number
        notes:
          type: string

    # ── SENSOR READING ───────────────────────────────────────────────────────────
    SensorReading:
      type: object
      required: [id, deviceId, soilMoisture, temperature, humidity, heatIndex, recordedAt]
      properties:
        id:
          type: integer
        deviceId:
          type: integer
        soilMoisture:
          type: number
          description: Soil moisture percentage (0-100)
        temperature:
          type: number
          description: Temperature in Celsius
        humidity:
          type: number
          description: Relative humidity percentage (0-100)
        heatIndex:
          type: number
          description: Computed heat index in Celsius
        electricalConductivity:
          type: ["number", "null"]
          description: Soil EC in mS/m (7-in-1 sensor)
        ph:
          type: ["number", "null"]
          description: Soil pH 0-14 (7-in-1 sensor)
        nitrogen:
          type: ["number", "null"]
          description: Available nitrogen in mg/kg (7-in-1 sensor)
        phosphorus:
          type: ["number", "null"]
          description: Available phosphorus in mg/kg (7-in-1 sensor)
        potassium:
          type: ["number", "null"]
          description: Available potassium in mg/kg (7-in-1 sensor)
        rainfall:
          type: ["number", "null"]
          description: Rainfall in mm
        lightIntensity:
          type: ["number", "null"]
          description: Light intensity in lux
        recordedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Alert:
      type: object
      required: [id, farmerId, type, severity, message, status, createdAt]
      properties:
        id:
          type: integer
        farmerId:
          type: ["integer", "null"]
        deviceId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [soil_moisture, temperature, humidity, pest, disease, irrigation, general]
        severity:
          type: string
          enum: [low, medium, high, critical]
        message:
          type: string
        status:
          type: string
          enum: [active, resolved, dismissed]
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        whatsappSent:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AlertInput:
      type: object
      required: [type, severity, message]
      properties:
        farmerId:
          type: integer
        deviceId:
          type: integer
        type:
          type: string
          enum: [soil_moisture, temperature, humidity, pest, disease, irrigation, general]
        severity:
          type: string
          enum: [low, medium, high, critical]
        message:
          type: string

    # ── RECOMMENDATION ───────────────────────────────────────────────────────────
...