# lib/api-zod/src/generated/types/farmerCreateResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AgroGuard Limited - Agricultural IoT Platform API
 * OpenAPI spec version: 0.2.0
 */
import type { FarmerCreateResultStatus } from './farmerCreateResultStatus';

export interface FarmerCreateResult {
  id: number;
  name: string;
  email: string;
  phone: string;
  location: string;
  /** @nullable */
  farmName?: string | null;
  /** @nullable */
  farmSizeHectares?: number | null;
  /** @nullable */
  cropTypes?: string | null;
  status: FarmerCreateResultStatus;
  /** @nullable */
  whatsappNumber?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  fieldOfficerId?: number | null;
  createdAt: Date;
  /** One-time temporary password shown only at creation; the farmer must change it on first login. */
  tempPassword: string;
}